+++ pyinseq/demultiplex.py
@@ -3,13 +3,16 @@
 Demultiplexes a FASTQ file into multiple files by 5' end barcode.
 
 Output path includes the experiment and sample name:
-(pysinseq)/results/{experiment}/{sample}.fastq
+(pyinseq)/results/{experiment}/{sample}.fastq
 
 """
-
+import tqdm
 import logging
 import re
 import screed
+import sys
+import os
+from .utils import count_sequences
 
 logger = logging.getLogger("pyinseq")
 
@@ -45,29 +48,48 @@ def demultiplex_fastq(reads, samples_dict: dict, settings) -> int:
     """,
         re.VERBOSE,
     )
-    with screed.open(reads) as seq_file:
-        for read in seq_file:
-            m = re.search(pattern, read.sequence)
-            try:
-                barcode, chrom_seq = m.group(1), m.group(2)
-                if barcode in demultiplex_dict:
-                    read["trim"] = m.span(2)  # trim slice
-                    demultiplex_dict[barcode].append(read)
-                else:
-                    demultiplex_dict["other"].append(read)
-            except:
+
+    # Initiate progress bar
+    logger.info("Preparing to demultiplex reads")
+
+    p_bar = tqdm.tqdm(
+        total=count_sequences(reads),
+        desc=f"Demultiplexing Reads",
+        unit="reads",
+        leave=False,
+    )
+    for read in screed.open(reads):
+        p_bar.update()
+        m = re.search(pattern, read.sequence)
+        try:
+            barcode, chrom_seq = m.group(1), m.group(2)
+            if barcode in demultiplex_dict:
+                read["trim"] = m.span(2)  # trim slice
+                demultiplex_dict[barcode].append(read)
+            else:
                 demultiplex_dict["other"].append(read)
-            # Every 10^6 sequences write and clear the dictionary
-            n_reads += 1
-            if n_reads % 5e6 == 0:
-                logger.info("Demultiplexed {:,} samples".format(n_reads))
-                write_reads(demultiplex_dict, samples_dict, settings)
-                # Write trimmed reads only when needed
-                if settings.write_trimmed_reads:
-                    write_trimmed_reads(demultiplex_dict, samples_dict, settings)
-                # Clear the dictionary after writing to file
-                for sample_name in demultiplex_dict:
-                    demultiplex_dict[sample_name] = []
+        except:
+            demultiplex_dict["other"].append(read)
+        # Every 10^6 sequences write and clear the dictionary
+        n_reads += 1
+        if n_reads % 5e6 == 0:
+            # Disable Progress Bar
+            p_bar.close()
+            logger.info(
+                "Demultiplexed {:,} samples and writing trimmed reads".format(n_reads)
+            )
+            # Enable Progress Bar
+            p_bar.disable = False
+            p_bar.refresh()
+            write_reads(demultiplex_dict, samples_dict, settings)
+            # Write trimmed reads only when needed
+            if settings.write_trimmed_reads:
+                write_trimmed_reads(demultiplex_dict, samples_dict, settings)
+            # Clear the dictionary after writing to file
+            for sample_name in demultiplex_dict:
+                demultiplex_dict[sample_name] = []
+    # Remove any traces of Progress Bar
+    p_bar.close()
     write_reads(demultiplex_dict, samples_dict, settings)
     # Write trimmed reads only when needed
     if settings.write_trimmed_reads:
@@ -87,8 +109,22 @@ def write_reads(demultiplex_dict, samples_dict, settings):
             with open(
                 f"{settings.path}raw_data/{barcode_dict[barcode]}.fastq", "a"
             ) as fo:
+                logger.debug(f"Writing reads to {barcode_dict[barcode]}.fastq")
+                # Initialize Progress Bar
+                p_bar = tqdm.tqdm(
+                    demultiplex_dict[barcode],
+                    total=len(demultiplex_dict[barcode]),
+                    desc=f"Writing {barcode_dict[barcode]} reads",
+                    unit="reads",
+                    leave=False,
+                    position=1,
+                )
                 for read in demultiplex_dict[barcode]:
+                    p_bar.update()
                     fo.write(f"@{read.name}\n{read.sequence}\n+\n{read.quality}\n")
+                # Remove any traces of Progress Bar
+                p_bar.close()
+    return
 
 
 def write_trimmed_reads(demultiplex_dict, samples_dict, settings):
@@ -102,10 +138,24 @@ def write_trimmed_reads(demultiplex_dict, samples_dict, settings):
             with open(
                 f"{settings.path}/{barcode_dict[barcode]}_trimmed.fastq", "a"
             ) as fo:
+                logger.debug(f"Writing trimmed reads to {barcode_dict[barcode]}.fastq")
+                # Initialize Progress Bar
+                p_bar = tqdm.tqdm(
+                    demultiplex_dict[barcode],
+                    total=len(demultiplex_dict[barcode]),
+                    desc=f"Writing {barcode_dict[barcode]} trimmed reads",
+                    unit="reads",
+                    leave=False,
+                    position=1,
+                )
                 for read in demultiplex_dict[barcode]:
+                    p_bar.update()
                     fo.write(
                         f"@{read.name}\n{read.sequence[slice(read.trim[0], read.trim[1])]}\n+\n{read.quality[slice(read.trim[0], read.trim[1])]}\n"
                     )
+                # Remove any traces of Progress Bar
+                p_bar.close()
+    return
 
 
 def main():