+++ pyinseq/utils.py
@@ -1,15 +1,35 @@
 #!/usr/bin/env python3
 
-import os
 import logging
+import os
 import re
+import sys
+import io
+import subprocess
+import tqdm
+import gzip
+
+
+class TqdmStream(io.StringIO):
+    """Custom stream with logging compatibility for tqdm progress bars."""
+
+    def __init__(self):
+        # Call parent constructor for instance variables
+        io.StringIO.__init__(self)
+
+    def write(self, s):
+        # Redirect through tqdm.write and output to stderr
+        tqdm.tqdm.write(s, file=sys.stderr, end="")
+
 
 # This controls the stdout logging.
 logging.basicConfig(
     level=logging.INFO,
     format="%(asctime)s - %(levelname)s - %(module)s - %(message)s",
     datefmt="%Y-%m-%d %H:%M",
+    stream=TqdmStream(),
 )
+
 logger = logging.getLogger("pyinseq")
 
 
@@ -65,6 +85,34 @@ def convert_to_filename(sample_name):
     Allows only alphanumeric, dashes, underscores, unicode.
     """
     return re.sub(r"(?u)[^-\w]", "", sample_name.strip().replace(" ", "_"))
+
+
+def count_sequences(filename):
+    """
+    Returns number of sequences in given file for progress bar.
+
+    Uses subprocess calls for efficiency
+    """
+    try:
+        # Check if gz compressed
+        if filename.split(".")[-1] in ["gz"]:
+            # Build pipe for uncompressed file
+            pipe = subprocess.Popen(("gzcat", "-d", filename), stdout=subprocess.PIPE)
+            output = int(subprocess.check_output(["wc", "-l"], stdin=pipe.stdout))
+        # If fastq format
+        elif filename.split(".")[-1] in ["fastq"]:
+            output = int(
+                subprocess.check_output(
+                    ["wc", "-l", filename], stderr=subprocess.DEVNULL
+                )
+                .split()[0]
+                .decode("UTF-8")
+            )
+        return int(output / 4)
+    except subprocess.CalledProcessError as e:
+        # Catching error returns arbitrary value
+        logger.debug(f"{str(e)}: using arbitrary limit for progress bar")
+        return 10e10
 
 
 # ===== Start here ===== #