+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Only Python 3.6+ supported
 ### Added
 - `pyinseq genomeprep` command
-- Add T50 calculation
+- Added T50 calculation
+- Added progress bar for `demultiplex` function
 
 ## [0.2.0] - 2017-07-16
 ### Added